requirements.txt
Django>=4.2,<6
djangorestframework>=3.15
psycopg2-binary>=2.9
pypdf>=3.0
python-docx>=0.8.11
python-pptx>=0.6.21
pandas>=2.0
pytesseract>=0.3.10
opencv-python>=4.8
Pillow>=10.0
sentence-transformers>=2.2.2
transformers>=4.34.0
langchain>=0.1.0
langgraph>=0.1.0
neo4j>=5.12.0
networkx>=3.1
chromadb>=0.4.0
faiss-cpu>=1.7.4
qdrant-client>=1.2.0
reportlab>=4.0
